+++ api/src/controllers/internal/Session.ts
@@ -97,6 +97,8 @@ ${autoExecContent}`
       session.path,
       '-AUTOEXEC',
       autoExecPath,
+      '-ENCODING',
+      'UTF-8',
       process.platform === 'win32' ? '-nosplash' : ''
     ])
       .then(() => {